refactor: move featureflag to core app
