+++ core/tests.py
@@ -8,7 +8,8 @@ from rest_framework import status
 from rest_framework.test import APITestCase
 
 from core.checks import REQUIRED_PROD_SETTINGS, check_required_prod_settings
-from core.models import GameSettings
+from core.admin import FeatureFlagForm
+from core.models import FeatureFlag, GameSettings
 from users.services.login_services import (
     calculate_daily_login_reward,
     LOGIN_STATE_ALREADY_LOGGED_TODAY,
@@ -261,3 +262,39 @@ class RequiredProdSettingsCheckTest(SimpleTestCase):
         with override_settings(**overrides):
             errors = check_required_prod_settings(app_configs=None)
         self.assertEqual(errors[0].id, "core.E001")
+
+
+_FAKE_FEATURE_FLAGS_TS = """
+const featureFlags = {
+  activityList: ['all'],
+  tasksFeature: ['testers'],
+  categoriesFeature: [],
+  skillsFeature: [],
+  projectsFeature: [],
+};
+"""
+
+
+@patch("core.admin._FEATURE_FLAGS_TS")
+class FeatureFlagFormKeyChoicesTests(TestCase):
+    def test_excludes_keys_already_created(self, mock_path):
+        mock_path.read_text.return_value = _FAKE_FEATURE_FLAGS_TS
+        FeatureFlag.objects.create(key="tasksFeature", access_groups=["testers"])
+
+        form = FeatureFlagForm()
+
+        choice_keys = [key for key, _ in form.fields["key"].choices]
+        self.assertNotIn("tasksFeature", choice_keys)
+        self.assertIn("activityList", choice_keys)
+        self.assertIn("categoriesFeature", choice_keys)
+
+    def test_editing_existing_flag_keeps_its_own_key_selectable(self, mock_path):
+        mock_path.read_text.return_value = _FAKE_FEATURE_FLAGS_TS
+        flag = FeatureFlag.objects.create(key="tasksFeature", access_groups=["testers"])
+        FeatureFlag.objects.create(key="activityList", access_groups=["all"])
+
+        form = FeatureFlagForm(instance=flag)
+
+        choice_keys = [key for key, _ in form.fields["key"].choices]
+        self.assertIn("tasksFeature", choice_keys)
+        self.assertNotIn("activityList", choice_keys)
